requirements-dev.txt
chardet==4.0.0
tox==3.25.0
Cython==0.29.28
# pylint==2.6.0
# wheel==0.35.1
# twine==3.2.0
# autopep8==1.5.4
